Update
The second "Main Setup" splits the implementation into different files
and has some general cleanup. This commit is to account for the
description I meant to add in the last one. The bulk of the files for
each commit is CMake files, which I don’t think has too much of an
impact.